Organize Services folders
Moved Auth, Users and Test services into separate folders
