+++ EUniversity/Extensions/ServiceCollectionExtensions.cs
@@ -1,13 +1,16 @@
 ﻿using EUniversity.Core.Models;
 using EUniversity.Core.Policy;
-using EUniversity.Core.Services;
+using EUniversity.Core.Services.Auth;
 using EUniversity.Core.Services.University;
+using EUniversity.Core.Services.Users;
 using EUniversity.Core.Validation.Auth;
 using EUniversity.Infrastructure.Data;
 using EUniversity.Infrastructure.Identity;
-using EUniversity.Infrastructure.Services;
+using EUniversity.Infrastructure.Services.Auth;
+using EUniversity.Infrastructure.Services.Test;
 using EUniversity.Infrastructure.Services.University;
 using EUniversity.Infrastructure.Services.University.Grades;
+using EUniversity.Infrastructure.Services.Users;
 using FluentValidation;
 using IdentityModel;
 using Microsoft.AspNetCore.Authentication;
